experiments/poc_4_rivers_extended/rag_experiment/requirements.txt
torch>=2.0.0
transformers>=4.30.0
numpy>=1.24.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
tqdm>=4.65.0
python-dotenv>=1.0.0
requests>=2.31.0
